Fix for port bindinging not being retriggered after sg set update on an instance
Updating the sg set on an instance causes the instance to be reassigned
to a new dvportgroup and a new port. The new port has a vlan 0 by
default which fails the check here and results in a stale model. The
port is then skipped from binding due to the stale model.