Reduce elements in the config
The number of xml elements is a limitating factor in the scale, especially when using the defaultPortConfig,
as each attribute will be multiplied by the number of number of ports, when updating the esxi hosts.

The destination of ingress (or source for egress) can be left blank. The protocol (IpV4/V6) will be determined by the respective reverse address.
A single ip is more efficiently expressed with SingleIp than a range.