Use a pool for portgroup-update requests against the vCenter
The vCenter cannot handle more than 16 tasks in parallel. If we have a
lot of updates all the time and just create tasks in the vCenter, the
vCenter cannot run them all and will start hanging.

Therefore, we re-use the pool passed to other classes for vCenter
operations, which don't seem to do much (anything?) with it. This
effectively limits the change-requests per dvs-agent.

This patch brought back a lot of stability in eu-de-2b.